+++ src/main/java/DBook/project/app/users/User.java
@@ -76,7 +76,7 @@ public class User implements Transactionable {
         }
     }
 
-    public Float calculateInvoice(Integer invoiceID, Transaction txt){ // pytanie, czy tu nie robimy opcji tylko dla swoich zamowien/offert (bo pracownik potrzebuje miec dostep do wszystkich)
+    public Double calculateInvoice(Integer invoiceID, Transaction txt){ // pytanie, czy tu nie robimy opcji tylko dla swoich zamowien/offert (bo pracownik potrzebuje miec dostep do wszystkich)
         return this.usersInvoices.get(invoiceID).calculateInvoice(txt);
     }
